Separate SimplePythonTransformer into separate subpackage
